backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy==1.4.49
psycopg2-binary>=2.9
python-dotenv>=1.0.0
pydantic>=2.0